Refactor TooltipBubbleShape
